+++ tests/test_source__tukui.py
@@ -1,11 +1,14 @@
 from __future__ import annotations
 
+from contextlib import nullcontext
+
 import pytest
 
 from instawow._sources.tukui import TukuiResolver
 from instawow.definitions import Defn
 from instawow.manager_ctx import ManagerCtx
 from instawow.pkg_models import Pkg
+from instawow.results import PkgFilesNotMatching
 from instawow.wow_installations import Flavour
 
 
@@ -23,15 +26,21 @@ def tukui_resolver(
 )
 @pytest.mark.parametrize('alias', ['tukui', 'elvui'])
 async def test_resolve_addon(
+    iw_manager_ctx: ManagerCtx,
     tukui_resolver: TukuiResolver,
     alias: str,
 ):
     defn = Defn('tukui', alias)
 
-    result = await tukui_resolver.resolve_one(defn, None)
+    with (
+        pytest.raises(PkgFilesNotMatching)
+        if iw_manager_ctx.config.game_flavour is Flavour.CataclysmClassic
+        else nullcontext()
+    ):
+        result = await tukui_resolver.resolve_one(defn, None)
 
-    assert type(result) is Pkg
-    assert result.slug == alias
+        assert type(result) is Pkg
+        assert result.slug == alias
 
 
 async def test_changelog_url_format(